+++ tests/e2e/support/assets.js
@@ -3,7 +3,7 @@ const path = require('path');
 const { WEBROOT, WEBROOT_FIRMWARE, FIRMWARE_PATH } = require('./paths');
 
 function hasNickelMenuAssets() {
-    return fs.existsSync(path.join(WEBROOT, 'assets', 'NickelMenu.zip')) && fs.existsSync(path.join(WEBROOT, 'assets', '.cog.png'));
+    return fs.existsSync(path.join(WEBROOT, 'assets', 'NickelMenu.tgz')) && fs.existsSync(path.join(WEBROOT, 'assets', '.cog.png'));
 }
 
 function hasKOReaderAssets() {