+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 # Changelog
 
 > This changelog is for the upcoming release. When a release is tagged, the list can be emptied.
+
+- Home-screen tweaks now install the standalone **NickelHome** mod alongside the standard NickelMenu, instead of relying on a NickelMenu fork. If your Kobo was set up with an older version of this tool, reinstalling updates NickelMenu to the standard version.
+- Removed the "NickelMenu only" (barebones) install option. Installing now always sets up the curated preset, which you can still fully customize.